Implement s3 emulator for testing

+++ src/megatron/energon/epathlib/epath.py
@@ -133,17 +133,21 @@ class EPath:
             target: The path to the local file to download to.
         """
 
-        if self.fs == target.fs:
-            self.fs.copy(self._internal_str_path, target._internal_str_path)
-        elif target.is_local():
-            # Note, src and target are swapped for these calls.
-            self.fs.download_file(target._internal_str_path, self._internal_str_path)
-        elif self.is_local():
-            # Note, src and target are swapped for these calls.
-            self.fs.upload_file(target._internal_str_path, self._internal_str_path)
+        if self.is_file():
+            if self.fs == target.fs:
+                self.fs.copy(self._internal_str_path, target._internal_str_path)
+            elif target.is_local():
+                self.fs.download_file(self._internal_str_path, target._internal_str_path)
+            elif self.is_local():
+                target.fs.upload_file(target._internal_str_path, self._internal_str_path)
+            else:
+                with self.open("rb") as src_f, target.open("wb") as dst_f:
+                    shutil.copyfileobj(src_f, dst_f)
         else:
-            with self.open("rb") as src_f, target.open("wb") as dst_f:
-                shutil.copyfileobj(src_f, dst_f)
+            inner_path = EPath(self)
+            for fpath in self.fs.list(self._internal_str_path):
+                inner_path.internal_path = PurePosixPath("/" + fpath.key)
+                inner_path.copy(target / inner_path.relative_to(self))
 
     @property
     def name(self) -> str:
